agent/requirements.txt
mss>=9.0.1
pyautogui>=0.9.54
Pillow>=10.0.0
pywinctl>=0.4
keyboard>=0.13.5
uiautomation>=2.0.18
pyperclip>=1.8.2
pytesseract>=0.3.10
